requirements.txt
# python 3.6
torch==1.8
pytorch-ignite==0.4.10
monai[itk, tensorboard, pytorch-ignite, torchvision]==0.8.1
scikit-image==0.17.2